requirements.txt
albumentations
numpy
opencv-python
pandas
pyyaml
scikit-learn
tensorflow

